requirements.txt
torch>=2.0
numpy>=1.24
scipy>=1.10
pandas>=2.0
PyYAML>=6.0
tqdm>=4.66
